requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.8
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
jsonify==0.5
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==1.5
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
urllib3==1.26.7
whitenoise==5.3.0
